% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyteGraph.R
\name{analyteGraph}
\alias{analyteGraph}
\title{analyteGraph}
\usage{
analyteGraph(model, stat = "fc", filter = "none", optns = list())
}
\arguments{
\item{model}{A model from PCA or oplsda function, or a data frame of analytes.}

\item{stat}{The statistic you would like to use. Either log2 fold change "fc",
Cliff's Delta "cd", p-value "pval". "fc", "pval" and "cd" are calculated
automatically. The p-value is the log 10 kruskal wallis p-value adjusted by
"bonferroni". You may select other p-value adjustments using method in the
optns.}

\item{filter}{Either "none", a range (1:10) or a numeric (1.3) for significance.
Based on internally calculated log 10 kruskal wallis p-value. A range will
provide only the top significant lipids (eg 1:10 will graph only top 10). A
numeric will filter out anything below this significance (eg 1.3, anything
with log 10 p-value less than 1.3 is excluded). "none" is the default and will
apply no filter.
\code{optns}.}

\item{optns}{A list for additional options:
\itemize{
\item{method} {adjustment for the internally calculated p-value. Default is
"bonferroni", but can be set to "holm", "hochberg", "hommel", "bonferroni",
"BH", "BY", "fdr" or "none".}
\item{lipidStart} {A character of the name of the first lipid in the range
you wish to exclude. If you want to exclude lipids, it is essential
to provide this and lipidEnd. Use the same syntax as you see in your model.}
\item{lipidEnd} {A character of the name of the last lipid in the range
you wish to exclude. Must be provided if lipidStart is provided.}
\item{discretePalette} {Color palette for discrete values, you can assign colors to
specific factors, example:
discretePalette = c("control" = "purple", "treatment" = "orange"). Or supply
a concatenated list, example (and the default):
discretePalette = c("#B2182B", "#D6604D", "#F4A582", "#FDDBC7", "#D1E5F0",
"#92C5DE", "#4393C3", "#2166AC"). Hexadecimal or color names accepted.}
\item{shape} {A character of the shape, Default shape is "circle".}
\item{alpha} {A numeric of the alpha desired. Default size is 0.3.}
\item{theme} {Personalize the plot theme you would like applied as you would
using theme() in ggplot. Example:
theme = theme(legend.position = "left", text=element_text(size=5)).}
\item{factor} {An object the same length as the data in the model supplied
that must have at least 2 unique groups such as treatment and control. More
than 2 groups is allowable.}
\item{control} {Character to set which group with the supplied factor you wish
to set as the control for comparison to all other groups.}
}}
}
\description{
Analyte versus statistic.
}
