% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binSpectra.R
\name{binSpectra}
\alias{binSpectra}
\title{Binning spectra}
\usage{
binSpectra(X, ppm, width = NULL, npoints = NULL)
}
\arguments{
\item{X}{full resolution spectra matrix}

\item{ppm}{parts per million}

\item{width}{interval in ppm. width must be larger than full resolution ppm interval}

\item{npoints}{total number of data points after binning npoints must be smaller than the length of ppm
You are required to specify either width or npoints NOT BOTH.}
}
\description{
Binning spectra
}
\examples{
Xb<-binSpectra(X = X, ppm = ppm, width = 0.001)
ppm_bin<-as.numeric(colnames(Xb))

}
