% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cliffsDelta.R
\name{cliffsDelta}
\alias{cliffsDelta}
\title{Cliff's Delta}
\usage{
cliffsDelta(model, optns = optns)
}
\arguments{
\item{model}{A PCA or oplsda object from mva.plots or a data frame passed to
this function.}

\item{optns}{A list passed to cliffsDelta with additional arguments.}
}
\description{
Cliff's Delta
}
