% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eruptionPlot.R
\name{eruptionPlot}
\alias{eruptionPlot}
\title{Eruption Plot.}
\usage{
eruptionPlot(model, optns = list())
}
\arguments{
\item{model}{A PCA or oplsda object.}

\item{optns}{A list for additional options:
\itemize{
\item{factor}{used when supplying a PCA object. An object the same
length as the data used to build the PCA model that must be a two
factor variable such as treatment and control.}
\item{color}{Color coding for the eruption plot, choose from the
adjusted and re-scaled p-value "pval", correlation "corr", log2 fold
change "fc" or cliff's delta "cd". The default is the correlation.}
\item{plotTitle}{A string specifying the plot title.Default is
"Eruption Plot".}
\item{method}{A string specifying the method parameter. Determines the
method to adjust p-values by. The options the same as listed in
stats::p.adjust ("holm", "hochberg", "hommel", "bonferroni", "BH",
"BY","fdr", "none"). The default is "bonferroni".}
\item{PC}{for a PCA object. A numeric for which principal component to
use for the loadings (for the plot y-axis) and scores (if correlation
is chosen for \code{colourCoding}).}
\item{continuousPalette}{Color palette
for continuous values, use hexadecimal values (example and default:
continuousPalette = c("#0000CC","#0000FF","#0055FF","#00AAFF","#00FFFF",
"#2BFFD5","#55FFAA","#80FF80","#AAFF55","#D4FF2B","#FFFF00","#FFAA00",
"#FF5500","#FF0000","#CC0000")),
grDevices names (example: continousPalette = rainbow(4)) or
color names (example : continuousPalette =c("purple", "orange")).}
\item{x}{Choose your x-axis using the same options stated for color.
The default is cliff's delta.}
\item{y}{Choose your y-axis using the same options stated for color.
The default is loadings.}
}}
}
\value{
The eruption plot is printed and the model is appended with the
eruption data (cliffs Delta, p-value, correlation, loadings). For ropls
object, eruptionData is found in suppLs.
}
\description{
Eruption plot.
}
\examples{
data(mtcars)

a <- PCA(data = mtcars[,1:7], center = TRUE, scale. = TRUE)
b <- eruptionPlot(model = a,
                  optns = list(factor = mtcars[,"vs"],
                              color = "corr",
                              plotTitle = "mtcars eruption",
                              method = "bonferroni"))
}
