% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foldChange.R
\name{foldChange}
\alias{foldChange}
\title{Fold Change}
\usage{
foldChange(model = model, optns = optns)
}
\arguments{
\item{model}{A PCA or oplsda object from mva.plots or a data frame passed to
this function.}

\item{optns}{A list passed to foldChange with additional arguments.}
}
\description{
Fold Change
}
