% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gPairsLower.R
\name{gPairsLower}
\alias{gPairsLower}
\title{Blank panel remover.}
\usage{
gPairsLower(g)
}
\arguments{
\item{g}{the ggpairs object (pcaGrid or plotLoadingsGrid plot).}
}
\description{
This function removes the blank squares (diagonal and upper half) from pcaGrid and plotLoadingsGrid.
}
