% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{lipidData}
\alias{lipidData}
\title{Lipid data stored as an RDS file.}
\format{
A data frame.
}
\usage{
data(lipidData)
}
\description{
This dataset contains lipid data.
}
\seealso{
\code{\link{readRDS}}
}
