% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lipidGraph.R
\name{lipidGraph}
\alias{lipidGraph}
\title{Lipid Graph}
\usage{
lipidGraph(model, stat = "fc", optns = list())
}
\arguments{
\item{model}{A model from PCA or oplsda function, or a dataframe of lipids.}

\item{stat}{The statistic you would like to use. Either log2 fold change "fc",
Cliff's Delta "cd", correlation "corr" or "external". "fc" and "cd" are
calculated automatically. "external" is an externally supplied statistic that
can be, for example, a p-value from lmm and needs to be provided in the
\code{optns}.}

\item{external}{If stat is set to external, set this parameter in the
\code{optns} list. Must be the same as the number of lipids.}

\item{factor}{A parameter for the \code{optns} list. An object the same
length as the data in the model supplied that must have at least 2 unique
groups such as treatment and control. More than 2 is allowable.}

\item{control}{A parameter for the \code{optns} list to set which group with
the supplied factor you wish to set as the control for comparison to all
other groups.}
}
\description{
Lipid plot of class vs side chain.
}
