% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manhattanPlot.R
\name{manhattanPlot}
\alias{manhattanPlot}
\title{Manhattan plot}
\usage{
manhattanPlot(
  data,
  x,
  y,
  point_shape = 1,
  point_size = 2,
  point_alpha = 1,
  show_legend = FALSE,
  log_scaling = TRUE,
  threshold = NA,
  xlab = "",
  ylab = "",
  main_title = "Manhattan Plot",
  interactive = FALSE,
  optns = list()
)
}
\arguments{
\item{data}{Data to be supplied}

\item{x}{numeric or factor, that goes on x-axis}

\item{y}{numeric, can be p-value, can be concentration depending on data}

\item{point_shape}{numeric, geom_point shape in ggplot. default is 1; an
empty circle}

\item{point_size}{numeric, geom_point size in ggplot. default is 2}

\item{point_alpha}{numeric, geom_point alpha in ggplot. default is 1}

\item{show_legend}{Boolean, to show legends on the right hand side of the
plot. default is FALSE. If interactive is TRUE then it overwrite to
show_legend = TRUE}

\item{log_scaling}{Boolean, to apply -log10 to the y value. default is TRUE.
If it is TRUE then threshold value is automatically log scaled.}

\item{threshold}{numeric, add horizontal dash line to the plot. If
log_scaling = TRUE then threshold value is automatically log scaled.}

\item{xlab}{The x-axis title.}

\item{ylab}{The y-axis title.}

\item{main_title}{A character argument for the title of the plot.}

\item{interactive}{Boolean, to make the plot interactive. default is FALSE.}

\item{optns}{An empty list for additional options.}
}
\value{
a ggplot2 object.
}
\description{
Manhattan plot
}
