% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matspec.R
\name{matspec}
\alias{matspec}
\title{matspec}
\usage{
matspec(
  X,
  ppm,
  roi = c(0.5, 9.5),
  interactive = TRUE,
  resolution = "full",
  ...
)
}
\arguments{
\item{X}{is a spectra matrix (colnames(X) <- ppm, and
rownames(X) <- Annotation$sampleID is recommended to do so before plotting)}

\item{ppm}{must be numeric and same length as the number of column in X}

\item{roi}{region of interest if not defined default is set as c(0.0 , 9.5)}

\item{interactive}{default is True which allow you to zoom in and select
which spectra to show}

\item{resolution}{character argument for resolution of non-interactive plots.
The default is "full", "partial" will result in lower resolution but faster
plotting.}

\item{...}{extra arguments.}
}
\value{
Interactive/non-interactive spectra plot
}
\description{
Plot multiple spectra plot that can be interactive (default)
}
\examples{
 \dontrun{
data("NMR_1D_rat_urine_spectra")
#use first 10 samples, note the spectra is only the region 2 to 4 ppm
Xsample <- X[1:10,]
matspec(X = Xsample,
        ppm = ppm,
        interactive = T)

# If using a dae
#nmr<-local(get(load("path to dae")))
#ppm<-as.numeric(nmr@varName)
#X<-nmr@.Data
#the first three samples
#matspec(X[1:3,], ppm, roi = c(3.0,4.5), interactive = T)
}
}
