% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mva.plots-package.R
\docType{package}
\name{mva.plots-package}
\alias{mva.plots}
\alias{mva.plots-package}
\title{mva.plots: Add visualisation capabilities to multivariate analysis}
\description{
What the package does (one paragraph). Multivariate plots.
}
\author{
\strong{Maintainer}: Lucy Grigoroff \email{Lucy.Grigoroff@murdoch.edu.au} (\href{https://orcid.org/o}{ORCID})

}
\keyword{internal}
