% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myFn1.R
\name{myFn1}
\alias{myFn1}
\title{Function 1 for ggpairs.}
\usage{
myFn1(data, mapping, method = "stat_ellipse")
}
\description{
Shell function for scatter plot of plotscores with ellipses for the pcaGrid function. Any aesthetic parameters are included in the pcaGrid function.
}
