% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{new_lipidMetadata}
\alias{new_lipidMetadata}
\title{Lipid metadata stored as an RDS file.}
\format{
A data frame.
}
\usage{
data(new_lipidMetadata)
}
\description{
This dataset contains metadata for lipid data.
}
\seealso{
\code{\link{readRDS}}
}
