% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oplsdaPredict.R
\name{oplsdaPredict}
\alias{oplsdaPredict}
\title{oplsdaPredict}
\usage{
oplsdaPredict(model, newdata, optns = list())
}
\arguments{
\item{model}{oplsda model (or ropls built model).}

\item{newdata}{Matrix or data frame of the same X variables as the model
(O-PLS-DA).}

\item{optns}{An empty list for confusion matrix addition.
\itemize{
\item{real} {The "real" classifications for the newdata as a factor. If
supplied, a confusion matrix will be calculated. Only available for DA.}
\item{orthoI} {If you have an OPLS with more than one orthogonal
component, you can specify how many you would like used in the
predictive calculation. If not specified, all available orthogonal
components will be used.}
}}
}
\value{
The prediction model including the predictive scores and orthogonal
scores.
}
\description{
Function that uses ropls::opls model to make predictions on a new data set.
Only the first opls predicitve and orthogonal component used currently.
}
\examples{
data(mtcars)
a <- oplsda(X = mtcars[,1:5],
            Y = mtcars[,8],
            type = "OPLS",
            optns = list(permI = 50))

jitterCol <- c("mpg", "cyl", "disp", "hp", "drat")

egData <- mtcars

egData[, jitterCol] <- lapply(X = egData[, jitterCol],
                              FUN = function(x) jitter(x, factor = 0.1))

b <- oplsdaPredict(model = a,
                   newdata = egData[,1:5])
}
