% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutationPlot.R
\name{permutationPlot}
\alias{permutationPlot}
\title{Plot Scores.}
\usage{
permutationPlot(model, optns = list())
}
\arguments{
\item{model}{An oplsda (ropls) model.}

\item{optns}{An empty list for additional options.
\itemize{
\item{plotTitle}{A character for the title of the plot.}
\item{colorQ}{A character of the color for Q2 (example "pink" or "#FF5500").
Default color is "red".}
\item{colorR}{A character of the color for R2Y (example "blue" or "#FF5500").
Default color is "blue".}
\item{shape}{A character (example "circle") or the number (example 21)
of the shape desired . Default shape is "circle".}
\item{size}{A numeric of the size desired. Default size is 3.}
\item{alpha}{A numeric of the alpha desired for the permuted results.
Default size is 0.4, actual values are always opaque.}
}}
}
\value{
Permutation plot of Q2 and R2Y. The caption denoted pQ2 and pR2Y
which indicates how significantly different the real results are from the
permuted. Example, pQ2=(1+n(permutedQ2>realQ2))/n(permutations)
}
\description{
Permutation plot for oplsda (ropls) models.
}
\examples{
data(mtcars)

a <- oplsda(X = mtcars[,1:7],
            Y = mtcars[,8],
            type = "OPLS",
            optns = list(permI = 50))

permutationPlot(model = a,
                optns=list(shape="square",
                           colorQ = "#FF5500",
                           colorR = "pink",
                           size = 5))
}
