% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPRROC.R
\name{plotPRROC}
\alias{plotPRROC}
\title{PR and ROC curve}
\usage{
plotPRROC(
  model = NULL,
  x = NULL,
  y = NULL,
  PC = c(1, 2),
  col = c("red", "black"),
  caption = FALSE
)
}
\arguments{
\item{model}{model produced by ropls opls() or mva-plots oplsda().}

\item{x, y}{are two vectors that contain elements for a class each. Class 1
and class 2 for x and y, respectively, are the default. If more that 2
classes are present, consider the One-vs-Rest scheme, which compares each
class against all the others (assumed as one) or the One-vs-One scheme, which
compares every unique pairwise combination of classes. For the One-vs-One,
you can specify different combinations of classes as x and y.}

\item{PC}{A concatenation of the principal components to use, default is c(1, 2)}

\item{col}{colors of the ROC and PR curves, supplied as name (eg "red") or
hexadecimal or color. Default it col = c("red", "black").}

\item{caption}{Boolean for AU ROC and AUC PR. If TRUE, they appear in the
center of the graph, if FALSE they appear as a caption at the botom of the
graph.}
}
\value{
PRROC_plot
}
\description{
Precision Recall (PR) and Receiver Operator Curves. Should be applied to
classifications problems, for example PLS-DA or O-PLS-DA.You can supply a
model (from ropls or mva plots if PLS-DA or OPLS-DA) or individual input of
x and y.
}
\examples{
data("mtcars")
Y<-as.factor(mtcars$vs)  # for DA

#PLS-DA
plsda_model<-oplsda(X = mtcars[,1:5], Y = Y, type = "PLS")
ropls_plsda_model<-ropls::opls(x = mtcars[,1:5], y = Y, predI = NA, orthoI = 0)

#O-PLS-DA
oplsda_model<-oplsda(X = mtcars[,1:5], Y = Y, type = "OPLS")
ropls_oplsda_model<-ropls::opls(x = mtcars[,1:5], y = Y, predI = 1, orthoI = NA)

t<-plotPRROC(oplsda_model, caption = FALSE)

#x and y
plotPRROC(x = oplsda_model@scoreMN[which(Y == 1)], y = oplsda_model@scoreMN[which(Y == 2)])
}
