% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stocsy.R
\name{plotStocsy}
\alias{plotStocsy}
\title{Simple STOCSY visualization}
\usage{
plotStocsy(aDF, breaks = 10)
}
\arguments{
\item{aDF, }{list, STOCSY results as returned by mvaPlots::stocsy()}

\item{breaks}{numeric, number of breaks in the x axis}
}
\value{
a ggplot of the STOCSY, with ppm in the horizontal axis, covariance
in the vertical axis, and correlation in the color scale
}
\description{
Simple STOCSY visualization
}
