% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specOverlay.R
\name{specOverlay}
\alias{specOverlay}
\title{specOverlay}
\usage{
specOverlay(
  X,
  ppm = NULL,
  roi = c(-0.01, 0.01),
  alp = 0.7,
  size = 0.5,
  title = "",
  optns = list(),
  ...
)
}
\arguments{
\item{X}{is a spectra matrix (colnames(X) <- ppm, and rownames(X)<-Annotation$sampleID is recommended to do so before plotting)}

\item{ppm}{must be numeric and same length as the number of column in X}

\item{roi}{region of interest if not defined default is set as c(-0.01 , 0.01)}

\item{size}{A numeric of the size desired. Default size is 0.5.}

\item{title}{A character for the title of the plot.}

\item{optns}{An empty list the first list represent facet (require), the second list as color (optional) and the third as linshape (optional).}

\item{alpha}{A numeric of the alpha desired. Default size is 0.7.}
}
\value{
matspec non-interactive version separated by assigned column factor (i.e., Groups) and within the each plot that can have different color or linetype to differentiate further by sub-categories (i.e., timepoints, class)
}
\description{
Plot multiple spectra plot that can be separated by groups, colour and line types (not interactive)
}
