% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symmetricLimits.R
\name{symmetric_limits}
\alias{symmetric_limits}
\title{symmetric_limits for ggplot object}
\usage{
symmetric_limits(x)
}
\arguments{
\item{x}{x coordinates.}
}
\description{
replaces the need for ggpmisc package
}
