% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myFn2.R
\name{symmetricLimits}
\alias{symmetricLimits}
\title{Function 2 for ggpairs.}
\usage{
symmetricLimits(x)
}
\arguments{
\item{x}{x coordinates used to map where ellipses are placed}
}
\description{
Shell function for loadings with text points for the plotLoadingsGrid
function. Any aesthetic parameters are included in the plotlLoadingsGrid
function. Alternative function for symmetric_limits from the ggpmisc package.
}
\keyword{internal}
