% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cliffsDelta.R
\name{cliffsDelta}
\alias{cliffsDelta}
\title{Cliff's Delta}
\usage{
cliffsDelta(model, optns = optns)
}
\arguments{
\item{model}{A PCA or oplsda object from mva.plots or a data frame.}

\item{optns}{A list passed to cliffsDelta with additional arguments.
\describ{
\item{factor} {An object the same length as the data in the model supplied
that must have at least 2 unique groups such as treatment and control. More
than 2 is allowable.}
\item{control} {Character to set which group with the supplied factor you wish
to set as the control for comparison to all other groups. If your factor is
already numeric, when specifying the control, supply it as a character, for
example, control = "0". If not manually set, the unique factor assigned one
from the supplied factor will be automatically selected}
}}
}
\value{
Creates a dataframe of Cliff's Detlas. For more than 2 groups, pairwise
calculations to the selected control will be returned.
}
\description{
Cliff's delta, a measure of a non-parametric effect size measure used to
quantify the difference between two groups or conditions. It provides a
measure of the ordinal association between two variables. Ranges between (-1)
and (1).
}
