% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matspec.R
\name{matspec}
\alias{matspec}
\title{matspec}
\usage{
matspec(X, ppm, roi = c(0.5, 9.5), interactive = TRUE, ...)
}
\arguments{
\item{X}{is a spectra matrix (colnames(X) <- ppm, and rownames(X)<-Annotation$sampleID is recommended to do so before plotting)}

\item{ppm}{must be numeric and same length as the number of column in X}

\item{roi}{region of interest if not defined default is set as c(0.0 , 9.5)}

\item{interactive}{default is True which allow you to zoom in and select which spectra to show}
}
\value{
Interactive/non-interactive spectra plot
}
\description{
Plot multiple spectra plot that can be interactive (default)
}
\examples{
# this cannot be we don't publish path to our data !nmr<-local(get(load("~/OneDrive - Murdoch University/datasets/Colchicin/DataElements/hims_colchicin_PLA_HIMr02_PROF.PLASMA.CPMG.daE")))
#ppm<-as.numeric(nmr@varName)
#X<-nmr@.Data
#matspec(X[1:3,],ppm,roi = c(3.0,4.5),interactive = T)
}
