% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mva.plots-package.R
\docType{package}
\name{mva.plots-package}
\alias{mva.plots}
\alias{mva.plots-package}
\title{mva.plots: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph). Multivariate plots.
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
