% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotLoadSpec.R
\name{PlotLoadSpec}
\alias{PlotLoadSpec}
\title{PlotLoadSpec}
\usage{
PlotLoadSpec(
  model,
  PC = 1,
  roi = c(0.5, 9.5),
  type = "Backscaled",
  X = NULL,
  option = list()
)
}
\arguments{
\item{model}{are the output from : PCA(), prcomp(), oplsda() or ropls::opls()}

\item{PC}{is the number of components that you want to plot the loadings of
Default is 1 and only require to change when plotting PCA or PLS model}

\item{roi}{is the region of interest (default is from 0.5 to 9.5 ppm)}

\item{type}{is the loading visualization types
"Statistical reconstruction" calculate the covariance for the trace
(y) and Pearson's correlation of the predictive (O)PLS scores with
each ppm for the color.//
"Backscaling" calculate the trace with predicted loadings multiplied
by the standard deviation of X feature, color is weighted by the
OPLS model.//
for PCA it is always "Statistical reconstruction".
but for PLS and OPLS "Backscaling" method is recommended.}

\item{X}{spectral data matrix used for the model.
This is needed for PCA with prcomp package, or
(O)PLS with type = "Statistical reconstruction".}
}
\value{
plot of the loadings
}
\description{
Allow to plot loadings of PCA, PLS, OPLS,PLS-DA and OPLS-DA
}
\examples{
# PCA model,the loading will always use type = "Statistical reconstruction".
#PCA_pca_model <- PCA(NOESY, annotation = ANN$COVID)  # PCA
#prcomp_pca_model<-prcomp(NOESY,scale. = T,rank. = 2) # PCA

#Y <- ANN$COVID
#Y<-as.numeric(gsub("preCOVID",1,gsub("COVID-pos",2,Y))) for (pls-da)
#Y1<-as.factor(ANN$COVID)  # for OPLS-DA
#ropls_pls_model<-ropls::opls(NOESY,y = Y)
#ropls_plsda_model<-ropls::opls(NOESY,y = ANN$COVID)
#ropls_opls_model<-ropls::opls(NOESY,y = Y,predI = 1, orthoI = NA)
#ropls_oplsda_model<-ropls::opls(NOESY,y = Y1,predI = 1, orthoI = NA)

#PlotLoadSpec(model = prcomp_pca_model)
}
