% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Storm.R
\name{storm}
\alias{storm}
\title{Subset Optimization by reference matching (STORM)}
\usage{
storm(
  X,
  ppm,
  b = 5,
  q = 0.01,
  idx_ref = NULL,
  roi = NULL,
  calibrate = FALSE,
  message = TRUE,
  plot = TRUE
)
}
\arguments{
\item{b}{expected half signal width data points}

\item{q}{predefined p-value threshold for refernce selection (0<q<=1, q<0.05 is significant).

Iteratively learns the true reference by repeating a procedure to find the most highly correlated spectra
and updating the reference multiple times.
It converges when a subset of samples has been found twice.}

\item{n}{expected number of spectra subset with hidden signal (signal of interest)}
}
\description{
Step 1 Subset selection
Step 2 STOCSY of subset
\itemize{
\item Pearson correlation of each subset spectrum with the reference (reference a.k.a driver, chemical shift)
\item calculate corresponding p-value by transforming the correlation into the t-statistics with n-2 degrees of freedoms
\item exclude spectra subsets that have negative correlation with the reference
\item from the remaining spectra subset with the lowest p-values are chosen to be in the subset
Step 3) reference updating
}
}
\references{
Posma et.al (2012) dx.doi.org/10.1021/ac302360v| Anal.Chem. 2012, 84, 10694−10701
}
