% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myFn2.R
\name{symmetricLimits}
\alias{symmetricLimits}
\title{Function 2 for ggpairs.}
\usage{
symmetricLimits(x)
}
\description{
Shell function for loadings with text points for the plotLoadingsGrid function. Any aesthetic parameters are included in the plotlLoadingsGrid function.
}
